Prevent issue of dotdrop not installing templates correctly when trash-cli is not installed on the user's system
Before, by just setting the trash_exe dyn variable to just 'which trash', if
trash wasn't installed on the user's system 'which trash' would exit with a
non-zero return status, which would cause dotdrop to error out.

I also took this as an opportunity to refactor how the 'editor_exe' dyn
variable is set.