Switch to invoking zgenom instead of zgen in .zshrc
This is in response to commit `eabdf55` in jandamm/zgenom, which no
longer automatically defines a 'zgen' shell function upon sourcing
'zgenom.zsh'. I could change the .zshrc to instead source 'zgen.zsh', or
I could decide to finally migrate the .zshrc as if I used zgenom the
entire time.